requirements.txt
langchain_core==1.2.28
langchain_openai==1.1.12
langgraph==1.1.6
loguru==0.7.3
pydantic==2.12.5
python-dotenv==1.2.2
